+++ src/cpp/build.sh
@@ -3,19 +3,32 @@
 set -e
 
 bdir="build"
-idir="install"
 do_install=false
 params=
 buildType="Debug"
 
 while [[ ! -z $1 ]]; do
   case $1 in
-  --clean) rm -rf $bdir $idir;;
+  --clean) rm -rf $bdir;;
   --release) buildTarget="Release";;
   --sysroot) SYSROOT_PATH="$2"; shift;;
   -i | --install) do_install=true;;
   +tests) params+=" -DENABLE_TESTS=ON";;
   +demo)  params+=" -DENABLE_DEMO_APP=ON";;
+  +gen-cov)
+    set -e
+    cd build
+    cp ../firebolt-open-rpc.json ./test/
+    ctest --test-dir ./test
+    mkdir -p coverage
+    gcovr -r .. \
+      --exclude '.*/test/.*\.h' \
+      --exclude '.*/test/.*\.cpp' \
+      --decisions \
+      --medium-threshold 50 --high-threshold 75 \
+      --html-details coverage/index.html \
+      --cobertura coverage.cobertura.xml
+    exit 0;;
   --) shift; break;;
   *) break;;
   esac; shift
@@ -31,7 +44,7 @@ if [[ ! -e $bdir ]]; then
     $params \
     "$@"
 fi
-cmake --build $bdir || exit $?
+cmake --build $bdir --parallel || exit $?
 if $do_install; then
   cmake --install $bdir || exit $?
 fi