+++ src/cpp/src/firebolt.cpp
@@ -17,16 +17,15 @@
  * limitations under the License.
  */
 
-#include "firebolt.h"
+#include "firebolt/firebolt.h"
 #include "accessibility_impl.h"
 #include "advertising_impl.h"
 #include "device_impl.h"
-#include "fireboltclient_version.h"
+#include "display_impl.h"
+#include "firebolt/client_version.h"
 #include "lifecycle_impl.h"
 #include "localization_impl.h"
-#include "metrics_impl.h"
 #include "presentation_impl.h"
-#include "securestorage_impl.h"
 #include "stats_impl.h"
 #include <firebolt/gateway.h>
 
@@ -39,11 +38,10 @@ public:
         : accessibility_(Firebolt::Helpers::GetHelperInstance()),
           advertising_(Firebolt::Helpers::GetHelperInstance()),
           device_(Firebolt::Helpers::GetHelperInstance()),
+          display_(Firebolt::Helpers::GetHelperInstance()),
           lifecycle_(Firebolt::Helpers::GetHelperInstance()),
           localization_(Firebolt::Helpers::GetHelperInstance()),
-          metrics_(Firebolt::Helpers::GetHelperInstance()),
           presentation_(Firebolt::Helpers::GetHelperInstance()),
-          secureStorage_(Firebolt::Helpers::GetHelperInstance()),
           stats_(Firebolt::Helpers::GetHelperInstance())
     {
     }
@@ -69,11 +67,10 @@ public:
     Accessibility::IAccessibility& AccessibilityInterface() override { return accessibility_; }
     Advertising::IAdvertising& AdvertisingInterface() override { return advertising_; }
     Device::IDevice& DeviceInterface() override { return device_; }
+    Display::IDisplay& DisplayInterface() override { return display_; }
     Lifecycle::ILifecycle& LifecycleInterface() override { return lifecycle_; }
     Localization::ILocalization& LocalizationInterface() override { return localization_; }
     Presentation::IPresentation& PresentationInterface() override { return presentation_; }
-    Metrics::IMetrics& MetricsInterface() override { return metrics_; }
-    SecureStorage::ISecureStorage& SecureStorageInterface() override { return secureStorage_; }
     Stats::IStats& StatsInterface() override { return stats_; }
 
 private:
@@ -89,11 +86,10 @@ private:
     Accessibility::AccessibilityImpl accessibility_;
     Advertising::AdvertisingImpl advertising_;
     Device::DeviceImpl device_;
+    Display::DisplayImpl display_;
     Lifecycle::LifecycleImpl lifecycle_;
     Localization::LocalizationImpl localization_;
-    Metrics::MetricsImpl metrics_;
     Presentation::PresentationImpl presentation_;
-    SecureStorage::SecureStorageImpl secureStorage_;
     Stats::StatsImpl stats_;
 };