+++ src/cpp/test/component/displayTest.cpp
@@ -1,0 +1,49 @@
+/*
+ * If not stated otherwise in this file or this component's LICENSE file the
+ * following copyright and licenses apply:
+ *
+ * Copyright 2025 Sky UK
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "firebolt/firebolt.h"
+#include "json_engine.h"
+#include <gtest/gtest.h>
+
+class DisplayTest : public ::testing::Test
+{
+protected:
+    JsonEngine jsonEngine;
+};
+
+TEST_F(DisplayTest, Size)
+{
+    auto expectedValue = jsonEngine.get_value("Display.size");
+    auto result = Firebolt::IFireboltAccessor::Instance().DisplayInterface().size();
+    ASSERT_TRUE(result) << "DisplayImpl::size() returned an error";
+
+    EXPECT_EQ(result->height, expectedValue.at("height").get<uint32_t>());
+    EXPECT_EQ(result->width, expectedValue.at("width").get<uint32_t>());
+}
+
+TEST_F(DisplayTest, MaxResolution)
+{
+    auto expectedValue = jsonEngine.get_value("Display.maxResolution");
+
+    auto result = Firebolt::IFireboltAccessor::Instance().DisplayInterface().maxResolution();
+    ASSERT_TRUE(result) << "DisplayImpl::maxResolution() returned an error";
+
+    EXPECT_EQ(result->height, expectedValue.at("height").get<uint32_t>());
+    EXPECT_EQ(result->width, expectedValue.at("width").get<uint32_t>());
+}